+++ src/Taxes/Contracts/TaxRate.php
@@ -19,6 +19,8 @@ use Vanilo\Contracts\Configurable;
 
 interface TaxRate extends Configurable
 {
+    public static function findOneForZone(Zone|int $zone, bool $activesOnly = true): ?TaxRate;
+
     public function getName(): string;
 
     public function getCalculator(): TaxCalculator;