+++ src/Taxes/Models/TaxRateProxy.php
@@ -16,6 +16,9 @@ namespace Vanilo\Taxes\Models;
 
 use Konekt\Concord\Proxies\ModelProxy;
 
+/**
+ * @method static \Vanilo\Taxes\Contracts\TaxRate|null findOneForZone(\Konekt\Address\Contracts\Zone|int $zone, bool $activesOnly = true)
+ */
 class TaxRateProxy extends ModelProxy
 {
 }