+++ src/Taxes/Providers/ModuleServiceProvider.php
@@ -15,11 +15,13 @@ declare(strict_types=1);
 namespace Vanilo\Taxes\Providers;
 
 use Konekt\Concord\BaseModuleServiceProvider;
+use Vanilo\Taxes\Calculators\NullTaxCalculator;
 use Vanilo\Taxes\Contracts\TaxRateResolver;
 use Vanilo\Taxes\Models\TaxCategory;
 use Vanilo\Taxes\Models\TaxCategoryType;
 use Vanilo\Taxes\Models\TaxRate;
 use Vanilo\Taxes\Resolver\TaxEngineManager;
+use Vanilo\Taxes\TaxCalculators;
 
 class ModuleServiceProvider extends BaseModuleServiceProvider
 {
@@ -38,5 +40,7 @@ class ModuleServiceProvider extends BaseModuleServiceProvider
 
         $this->app->singleton(TaxEngineManager::class, fn ($app) => new TaxEngineManager($app));
         $this->app->bind(TaxRateResolver::class, fn ($app) => $app->make(TaxEngineManager::class)->driver());
+
+        TaxCalculators::register('none', NullTaxCalculator::class);
     }
 }