+++ src/Taxes/Changelog.md
@@ -12,6 +12,7 @@
 - Changed minimum Laravel version to v10.38.2
 - Changed the minimal Enum requirement to v4.2
 - BC: Changed the `TaxRate` interface so that it extends the `Configurable` interface
+- BC: Added the `findOneForZone` static method to the `TaxRate` interface
 - Added the `type` field to the TaxCategory model
 - Added the `Taxable` interface
 - Added the extendable `TaxEngine` (facade) that can resolve tax rates from taxables, billing/shipping addresses (a place for various country-specific taxation drivers)